# refactor: use slices.Sort for better performance