… LICENSES/vendor/golang.org/x/exp/LICENSE
… go.sum
… vendor/golang.org/x/exp/LICENSE
… vendor/golang.org/x/exp/PATENTS
… vendor/golang.org/x/exp/constraints/constraints.go
… vendor/golang.org/x/exp/slices/slices.go
… vendor/golang.org/x/exp/slices/sort.go
… vendor/golang.org/x/exp/slices/zsortfunc.go
… vendor/golang.org/x/exp/slices/zsortordered.go
… vendor/modules.txt