@enj 2023-03-08
Please avoid these types of refactors, they cause a lot of review churn and do not have a measurable impact on end users.  Focus performance improvement efforts on the critical path.